Change of package name and new class (TribesConfig) for refactoring
