+++ src/core/TribesConfig.java
@@ -1,0 +1,17 @@
+package core;
+
+public class TribesConfig
+{
+
+    // Warrior
+    public static int WARRIOR_ATTACK = 2;
+    public static int WARRIOR_DEFENCE = 2;
+    public static int WARRIOR_MOVEMENT = 1;
+    public static int WARRIOR_MAX_HP = 10;
+    public static int WARRIOR_RANGE = 1;
+    public static int WARRIOR_COST = 2;
+
+    //Archer ...
+
+
+}
