+++ src/core/actions/cityactions/Destroy.java
@@ -1,6 +1,6 @@
 package core.actions.cityactions;
 
-import core.units.City;
+import core.actors.City;
 
 public class Destroy extends CityAction
 {
