+++ src/core/actions/cityactions/Spawn.java
@@ -1,7 +1,7 @@
 package core.actions.cityactions;
 
 import core.Types;
-import core.units.City;
+import core.actors.City;
 
 public class Spawn extends CityAction
 {
