+++ src/core/actions/unitactions/Capture.java
@@ -1,7 +1,7 @@
 package core.actions.unitactions;
 
-import core.units.City;
-import core.units.Unit;
+import core.actors.City;
+import core.actors.units.Unit;
 
 public class Capture extends UnitAction
 {
