+++ src/core/actors/buildings/Mine.java
@@ -1,4 +1,4 @@
-package core.units;
+package core.actors.buildings;
 
 import core.Types;
