+++ src/core/actors/buildings/MountainTemple.java
@@ -1,24 +1,23 @@
-package core.units;
+package core.actors.buildings;
 
 import core.Types;
 
-public class Temple extends Building {
+public class MountainTemple extends Building {
 
     private final Types.BUILDING TYPE = Types.BUILDING.TEMPLE;
     private final int COST = 20;
     private final int PRODUCTION = 1;
     private final int points = 100;
-    // TODO: It needs to check if the resource is field tile or not
-    private final Types.TERRAIN TERRAIN_CONSTRAINT = Types.TERRAIN.PLAIN;
+    private final Types.TERRAIN TERRAIN_CONSTRAINT = Types.TERRAIN.MOUNTAIN;
 
 
-    public Temple(int x, int y) {
+    public MountainTemple(int x, int y) {
         super(x, y);
     }
 
     @Override
     public Building copy() {
-        return new Temple(getX(), getY());
+        return new MountainTemple(getX(), getY());
     }
 
     @Override
