+++ src/core/actors/units/Knight.java
@@ -1,4 +1,4 @@
-package core.units;
+package core.actors.units;
 
 import utils.Vector2d;
