+++ src/core/actors/units/Warrior.java
@@ -1,11 +1,13 @@
-package core.units;
+package core.actors.units;
 
 import utils.Vector2d;
+
+import static core.TribesConfig.*;
 
 public class Warrior extends Unit
 {
     public Warrior(Vector2d pos, int kills, boolean isVeteran, int ownerID) {
-        super(2, 2, 1, 10, 1, 2, pos, kills, isVeteran, ownerID);
+        super(WARRIOR_ATTACK, WARRIOR_DEFENCE, WARRIOR_MOVEMENT, WARRIOR_MAX_HP, WARRIOR_RANGE, WARRIOR_COST, pos, kills, isVeteran, ownerID);
     }
 
     @Override
