+++ src/core/game/Game.java
@@ -2,7 +2,7 @@ package core.game;
 
 import core.Types;
 import core.actions.Action;
-import core.units.Tribe;
+import core.actors.Tribe;
 import players.Agent;
 import utils.ElapsedCpuTimer;
 import utils.GUI;
